uuid

Module adapter

Adapters for various formats for UUIDs

Structs

Hyphenated

An adaptor for formatting an Uuid as a hyphenated string.

HyphenatedRef

An adaptor for formatting an Uuid as a hyphenated string.

Simple

An adaptor for formatting an Uuid as a simple string.

SimpleRef

An adaptor for formatting an Uuid as a simple string.

Urn

An adaptor for formatting an Uuid as a URN string.

UrnRef

An adaptor for formatting an Uuid as a URN string.