Struct uuid::adapter::SimpleRef

pub struct SimpleRef<'a>(_);

An adaptor for formatting an Uuid as a simple string.

Takes a reference of the Uuid.

Implementations

impl<'a> SimpleRef<'a>

pub const LENGTH: usize

The length of a simple Uuid string.

pub const fn from_uuid_ref(uuid: &'a Uuid) -> Self

Creates a SimpleRef from a Uuid reference.

pub fn encode_lower<'buf>(&self, buffer: &'buf mut [u8]) -> &'buf mut str

Writes the Uuid as a lower-case simple string to buffer, and returns the subslice of the buffer that contains the encoded UUID.

This is slightly more efficient than using the formatting infrastructure as it avoids virtual calls, and may avoid double buffering.

Panics

Panics if the buffer is not large enough: it must have length at least LENGTH. Uuid::encode_buffer can be used to get a sufficiently-large temporary buffer.

Examples

use uuid::Uuid;

fn main() -> Result<(), uuid::Error> {
    let uuid = Uuid::parse_str("936DA01f9abd4d9d80c702af85c822a8")?;

    // the encoded portion is returned
    assert_eq!(
        uuid.to_simple().encode_lower(&mut Uuid::encode_buffer()),
        "936da01f9abd4d9d80c702af85c822a8"
    );

    // the buffer is mutated directly, and trailing contents remains
    let mut buf = [b'!'; 36];
    assert_eq!(
        uuid.to_simple().encode_lower(&mut buf),
        "936da01f9abd4d9d80c702af85c822a8"
    );
    assert_eq!(
        &buf as &[_],
        b"936da01f9abd4d9d80c702af85c822a8!!!!" as &[_]
    );

    Ok(())
}

*/

pub fn encode_upper<'buf>(&self, buffer: &'buf mut [u8]) -> &'buf mut str

Writes the Uuid as an upper-case simple string to buffer, and returns the subslice of the buffer that contains the encoded UUID.

Panics

Panics if the buffer is not large enough: it must have length at least LENGTH. Uuid::encode_buffer can be used to get a sufficiently-large temporary buffer.

Examples

use uuid::Uuid;

fn main() -> Result<(), uuid::Error> {
    let uuid = Uuid::parse_str("936da01f9abd4d9d80c702af85c822a8")?;

    // the encoded portion is returned
    assert_eq!(
        uuid.to_simple().encode_upper(&mut Uuid::encode_buffer()),
        "936DA01F9ABD4D9D80C702AF85C822A8"
    );

    // the buffer is mutated directly, and trailing contents remains
    let mut buf = [b'!'; 36];
    assert_eq!(
        uuid.to_simple().encode_upper(&mut buf),
        "936DA01F9ABD4D9D80C702AF85C822A8"
    );
    assert_eq!(
        &buf as &[_],
        b"936DA01F9ABD4D9D80C702AF85C822A8!!!!" as &[_]
    );

    Ok(())
}

*/