Struct uuid::adapter::UrnRef

pub struct UrnRef<'a>(_);

An adaptor for formatting an Uuid as a URN string.

Takes a reference of the Uuid.

Methods

impl<'a> UrnRef<'a>

pub const LENGTH: usize

The length of a URN Uuid string.

pub fn from_uuid_ref(uuid: &'a Uuid) -> Self

Creates a UrnRef from a Uuid reference.

pub fn encode_lower<'buf>(&self, buffer: &'buf mut [u8]) -> &'buf mut str

Writes the Uuid as a lower-case URN string to buffer, and returns the subslice of the buffer that contains the encoded UUID.

This is slightly more efficient than using the formatting infrastructure as it avoids virtual calls, and may avoid double buffering.

Panics

Panics if the buffer is not large enough: it must have length at least LENGTH. Uuid::encode_buffer can be used to get a sufficiently-large temporary buffer.

Examples

use uuid::Uuid;

let uuid = Uuid::parse_str("936DA01f9abd4d9d80c702af85c822a8").unwrap();

// the encoded portion is returned
assert_eq!(
    uuid.to_urn().encode_lower(&mut Uuid::encode_buffer()),
    "urn:uuid:936da01f-9abd-4d9d-80c7-02af85c822a8"
);

// the buffer is mutated directly, and trailing contents remains
let mut buf = [b'!'; 49];
uuid.to_urn().encode_lower(&mut buf);
assert_eq!(
    uuid.to_urn().encode_lower(&mut buf),
    "urn:uuid:936da01f-9abd-4d9d-80c7-02af85c822a8"
);
assert_eq!(
    &buf as &[_],
    b"urn:uuid:936da01f-9abd-4d9d-80c7-02af85c822a8!!!!" as &[_]
);

*/

pub fn encode_upper<'buf>(&self, buffer: &'buf mut [u8]) -> &'buf mut str

Writes the Uuid as an upper-case URN string to buffer, and returns the subslice of the buffer that contains the encoded UUID.

This is slightly more efficient than using the formatting infrastructure as it avoids virtual calls, and may avoid double buffering.

Panics

Panics if the buffer is not large enough: it must have length at least LENGTH. Uuid::encode_buffer can be used to get a sufficiently-large temporary buffer.

Examples

use uuid::Uuid;

let uuid = Uuid::parse_str("936da01f9abd4d9d80c702af85c822a8").unwrap();

// the encoded portion is returned
assert_eq!(
    uuid.to_urn().encode_upper(&mut Uuid::encode_buffer()),
    "urn:uuid:936DA01F-9ABD-4D9D-80C7-02AF85C822A8"
);

// the buffer is mutated directly, and trailing contents remains
let mut buf = [b'!'; 49];
assert_eq!(
    uuid.to_urn().encode_upper(&mut buf),
    "urn:uuid:936DA01F-9ABD-4D9D-80C7-02AF85C822A8"
);
assert_eq!(
    &buf as &[_],
    b"urn:uuid:936DA01F-9ABD-4D9D-80C7-02AF85C822A8!!!!" as &[_]
);

*/