uuid
Module error

Modules

std_support 🔒

Structs

Error

A general error that can occur when working with UUIDs.

InvalidUuid

A string that is guaranteed to fail to parse to a Uuid.

Enums

ErrorKind 🔒