uuid

Module fmt

Adapters for alternative string formats.

Macros

impl_fmt_from 🔒

impl_fmt_traits 🔒

Structs

Braced

Format a Uuid as a braced hyphenated string, like {67e55044-10b1-426f-9247-bb680e5fe0c8}.

Hyphenated

Format a Uuid as a hyphenated string, like 67e55044-10b1-426f-9247-bb680e5fe0c8.

Simple

Format a Uuid as a simple string, like 67e5504410b1426f9247bb680e5fe0c8.

Urn

Format a Uuid as a URN string, like urn:uuid:67e55044-10b1-426f-9247-bb680e5fe0c8.

Functions

assume_init_ascii_mut 🔒 ⚠

encode_braced 🔒

encode_braced_uninit 🔒

encode_hyphenated 🔒

encode_hyphenated_uninit 🔒

encode_simple 🔒

encode_simple_uninit 🔒

encode_urn 🔒

encode_urn_uninit 🔒

format_hyphenated 🔒

format_simple 🔒

nibble_to_hex 🔒

slice_as_uninit_mut 🔒

write_bytes 🔒