Struct uuid::fmt::Braced

#[repr(transparent)]
pub struct Braced(_);

Format a Uuid as a braced hyphenated string, like {67e55044-10b1-426f-9247-bb680e5fe0c8}.

Implementations

impl Braced

pub const LENGTH: usize = 38usize

The length of a braced Uuid string.

pub const fn from_uuid(uuid: Uuid) -> Self

Creates a Braced from a Uuid.

pub fn encode_lower<'buf>(&self, buffer: &'buf mut [u8]) -> &'buf mut str

Writes the Uuid as a lower-case hyphenated string surrounded by braces to buffer, and returns the subslice of the buffer that contains the encoded UUID.

This is slightly more efficient than using the formatting infrastructure as it avoids virtual calls, and may avoid double buffering.

Panics

Panics if the buffer is not large enough: it must have length at least LENGTH. Uuid::encode_buffer can be used to get a sufficiently-large temporary buffer.

Examples

use uuid::Uuid;

fn main() -> Result<(), uuid::Error> {
    let uuid = Uuid::parse_str("936DA01f9abd4d9d80c702af85c822a8")?;

    // the encoded portion is returned
    assert_eq!(
        uuid.braced()
            .encode_lower(&mut Uuid::encode_buffer()),
        "{936da01f-9abd-4d9d-80c7-02af85c822a8}"
    );

    // the buffer is mutated directly, and trailing contents remains
    let mut buf = [b'!'; 40];
    uuid.braced().encode_lower(&mut buf);
    assert_eq!(
        &buf as &[_],
        b"{936da01f-9abd-4d9d-80c7-02af85c822a8}!!" as &[_]
    );

    Ok(())
}

*/

pub fn encode_upper<'buf>(&self, buffer: &'buf mut [u8]) -> &'buf mut str

Writes the Uuid as an upper-case hyphenated string surrounded by braces to buffer, and returns the subslice of the buffer that contains the encoded UUID.

This is slightly more efficient than using the formatting infrastructure as it avoids virtual calls, and may avoid double buffering.

Panics

Panics if the buffer is not large enough: it must have length at least LENGTH. Uuid::encode_buffer can be used to get a sufficiently-large temporary buffer.

Examples

use uuid::Uuid;

fn main() -> Result<(), uuid::Error> {
    let uuid = Uuid::parse_str("936da01f9abd4d9d80c702af85c822a8")?;

    // the encoded portion is returned
    assert_eq!(
        uuid.braced()
            .encode_upper(&mut Uuid::encode_buffer()),
        "{936DA01F-9ABD-4D9D-80C7-02AF85C822A8}"
    );

    // the buffer is mutated directly, and trailing contents remains
    let mut buf = [b'!'; 40];
    uuid.braced().encode_upper(&mut buf);
    assert_eq!(
        &buf as &[_],
        b"{936DA01F-9ABD-4D9D-80C7-02AF85C822A8}!!" as &[_]
    );

    Ok(())
}

*/

pub const fn as_uuid(&self) -> &Uuid

Get a reference to the underlying Uuid.

Examples

use uuid::Uuid;

let braced = Uuid::nil().braced();
assert_eq!(*braced.as_uuid(), Uuid::nil());

pub const fn into_uuid(self) -> Uuid

Consumes the Braced, returning the underlying Uuid.

Examples

use uuid::Uuid;

let braced = Uuid::nil().braced();
assert_eq!(braced.into_uuid(), Uuid::nil());