Module uuid::prelude

The uuid prelude.

This module contains the most important items of the uuid crate.

To use the prelude, include the following in your crate root:

extern crate uuid;

Prelude Contents

Currently the prelude reexports the following:

uuid::{Error, Uuid, Variant, Version, builder::Builder}: The fundamental types used in uuid crate.

Re-exports

pub use super::Builder;

pub use super::Bytes;

pub use super::Error;

pub use super::Uuid;

pub use super::Variant;

pub use super::Version;