Struct uuid::prelude::Uuid

pub struct Uuid { /* fields omitted */ }

A Universally Unique Identifier (UUID).

Methods

impl Uuid

pub fn nil() -> Uuid

The 'nil UUID'.

The nil UUID is special form of UUID that is specified to have all 128 bits set to zero, as defined in IETF RFC 4122 Section 4.1.7.

Examples

Basic usage:

use uuid::Uuid;

let uuid = Uuid::nil();

assert_eq!(uuid.hyphenated().to_string(),
           "00000000-0000-0000-0000-000000000000");

pub fn new(v: UuidVersion) -> Option<Uuid>

Creates a new Uuid.

Note that not all versions can be generated currently and None will be returned if the specified version cannot be generated.

To generate a random UUID (UuidVersion::Md5), then the v3 feature must be enabled for this crate.

To generate a random UUID (UuidVersion::Random), then the v4 feature must be enabled for this crate.

To generate a random UUID (UuidVersion::Sha1), then the v5 feature must be enabled for this crate.

pub fn new_v4() -> Uuid

Creates a random Uuid.

This uses the rand crate's default task RNG as the source of random numbers. If you'd like to use a custom generator, don't use this method: use the rand::Rand trait's rand() method instead.

Note that usage of this method requires the v4 feature of this crate to be enabled.

Examples

Basic usage:

use uuid::Uuid;

let uuid = Uuid::new_v4();

pub fn from_fields(
d1: u32,
d2: u16,
d3: u16,
d4: &[u8]
) -> Result<Uuid, ParseError>

Creates a Uuid from four field values.

Errors

This function will return an error if d4's length is not 8 bytes.

Examples

Basic usage:

use uuid::Uuid;

let d4 = [12, 3, 9, 56, 54, 43, 8, 9];

let uuid = Uuid::from_fields(42, 12, 5, &d4);
let uuid = uuid.map(|uuid| uuid.hyphenated().to_string());

let expected_uuid = Ok(String::from("0000002a-000c-0005-0c03-0938362b0809"));

assert_eq!(expected_uuid, uuid);

An invalid length:

use uuid::Uuid;
use uuid::ParseError;

let d4 = [12];

let uuid = Uuid::from_fields(42, 12, 5, &d4);

let expected_uuid = Err(ParseError::InvalidLength(1));

assert_eq!(expected_uuid, uuid);

pub fn from_bytes(b: &[u8]) -> Result<Uuid, ParseError>

Creates a Uuid using the supplied bytes.

Errors

This function will return an error if b has any length other than 16.

Examples

Basic usage:

use uuid::Uuid;

let bytes = [4, 54, 67, 12, 43, 2, 98, 76,
             32, 50, 87, 5, 1, 33, 43, 87];

let uuid = Uuid::from_bytes(&bytes);
let uuid = uuid.map(|uuid| uuid.hyphenated().to_string());

let expected_uuid = Ok(String::from("0436430c-2b02-624c-2032-570501212b57"));

assert_eq!(expected_uuid, uuid);

An incorrect number of bytes:

use uuid::Uuid;
use uuid::ParseError;

let bytes = [4, 54, 67, 12, 43, 2, 98, 76];

let uuid = Uuid::from_bytes(&bytes);

let expected_uuid = Err(ParseError::InvalidLength(8));

assert_eq!(expected_uuid, uuid);

pub fn from_uuid_bytes(bytes: UuidBytes) -> Uuid

Creates a Uuid using the supplied bytes.

Examples

Basic usage:

use uuid::Uuid;
use uuid::UuidBytes;

let bytes:UuidBytes = [70, 235, 208, 238, 14, 109, 67, 201, 185, 13, 204, 195, 90, 145, 63, 62];

let uuid = Uuid::from_uuid_bytes(bytes);
let uuid = uuid.hyphenated().to_string();

let expected_uuid = String::from("46ebd0ee-0e6d-43c9-b90d-ccc35a913f3e");

assert_eq!(expected_uuid, uuid);

An incorrect number of bytes:

use uuid::Uuid;
use uuid::UuidBytes;

let bytes:UuidBytes = [4, 54, 67, 12, 43, 2, 98, 76]; // doesn't compile

let uuid = Uuid::from_uuid_bytes(bytes);

pub fn from_random_bytes(b: [u8; 16]) -> Uuid

Creates a v4 Uuid from random bytes (e.g. bytes supplied from Rand crate)

Examples

Basic usage:

use uuid::Uuid;
use uuid::UuidBytes;

let bytes:UuidBytes = [70, 235, 208, 238, 14, 109, 67, 201, 185, 13, 204, 195, 90, 145, 63, 62];
let uuid = Uuid::from_random_bytes(bytes);
let uuid = uuid.hyphenated().to_string();

let expected_uuid = String::from("46ebd0ee-0e6d-43c9-b90d-ccc35a913f3e");

assert_eq!(expected_uuid, uuid);

pub fn get_variant(&self) -> Option<UuidVariant>

Returns the variant of the Uuid structure.

This determines the interpretation of the structure of the UUID. Currently only the RFC4122 variant is generated by this module.

Variant Reference

pub fn get_version_num(&self) -> usize

Returns the version number of the Uuid.

This represents the algorithm used to generate the contents.

Currently only the Random (V4) algorithm is supported by this module. There are security and privacy implications for using older versions - see Wikipedia: Universally Unique Identifier for details.

Version Reference

pub fn get_version(&self) -> Option<UuidVersion>

Returns the version of the Uuid.

This represents the algorithm used to generate the contents

pub fn as_fields(&self) -> (u32, u16, u16, &[u8; 8])

Returns the four field values of the UUID.

These values can be passed to the from_fields() method to get the original Uuid back.

The first field value represents the first group of (eight) hex digits, taken as a big-endian u32 value. For V1 UUIDs, this field represents the low 32 bits of the timestamp.

The second field value represents the second group of (four) hex digits, taken as a big-endian u16 value. For V1 UUIDs, this field represents the middle 16 bits of the timestamp.

The third field value represents the third group of (four) hex digits, taken as a big-endian u16 value. The 4 most significant bits give the UUID version, and for V1 UUIDs, the last 12 bits represent the high 12 bits of the timestamp.

The last field value represents the last two groups of four and twelve hex digits, taken in order. The first 1-3 bits of this indicate the UUID variant, and for V1 UUIDs, the next 13-15 bits indicate the clock sequence and the last 48 bits indicate the node ID.

Examples

use uuid::Uuid;

let uuid = Uuid::nil();
assert_eq!(uuid.as_fields(), (0, 0, 0, &[0u8; 8]));

let uuid = Uuid::parse_str("936DA01F-9ABD-4D9D-80C7-02AF85C822A8").unwrap();
assert_eq!(uuid.as_fields(),
           (0x936DA01F, 0x9ABD, 0x4D9D, b"\x80\xC7\x02\xAF\x85\xC8\x22\xA8"));

pub fn as_bytes(&self) -> &[u8; 16]

Returns an array of 16 octets containing the UUID data.

Examples

use uuid::Uuid;

let uuid = Uuid::nil();
assert_eq!(uuid.as_bytes(), &[0; 16]);

let uuid = Uuid::parse_str("936DA01F9ABD4d9d80C702AF85C822A8").unwrap();
assert_eq!(uuid.as_bytes(),
           &[147, 109, 160, 31, 154, 189, 77, 157,
             128, 199, 2, 175, 133, 200, 34, 168]);

pub fn simple(&self) -> Simple

Returns a wrapper which when formatted via fmt::Display will format a string of 32 hexadecimal digits.

Examples

use uuid::Uuid;

let uuid = Uuid::nil();
assert_eq!(uuid.simple().to_string(),
           "00000000000000000000000000000000");

pub fn hyphenated(&self) -> Hyphenated

Returns a wrapper which when formatted via fmt::Display will format a string of hexadecimal digits separated into groups with a hyphen.

Examples

use uuid::Uuid;

let uuid = Uuid::nil();
assert_eq!(uuid.hyphenated().to_string(),
           "00000000-0000-0000-0000-000000000000");

pub fn urn(&self) -> Urn

Returns a wrapper which when formatted via fmt::Display will format a string of the UUID as a full URN string.

Examples

use uuid::Uuid;

let uuid = Uuid::nil();
assert_eq!(uuid.urn().to_string(),
           "urn:uuid:00000000-0000-0000-0000-000000000000");

pub fn to_timestamp(&self) -> Option<(u64, u16)>

Returns an Optional Tuple of (u64, u16) representing the timestamp and counter portion of a V1 UUID. If the supplied UUID is not V1, this will return None

pub fn parse_str(input: &str) -> Result<Uuid, ParseError>

Parses a Uuid from a string of hexadecimal digits with optional hyphens.

Any of the formats generated by this module (simple, hyphenated, urn) are supported by this parsing function.

pub fn is_nil(&self) -> bool

Tests if the UUID is nil