Module uuid::serde

Adapters for alternative serde formats.

This module contains adapters you can use with #[serde(with)] to change the way a Uuid is serialized and deserialized.

Modules

compact

Serialize a Uuid as a [u8; 16].