uuid::serde

Module braced

Serialize a Uuid as [uuid::fmt::Braced].

Examples

Serialize and deserialize using the braced format, failing to deserialize any other format:

#[derive(serde_derive::Serialize, serde_derive::Deserialize)]
struct StructA {
    #[serde(with = "uuid::serde::braced")]
    id: uuid::Uuid,
}

Serialize using the braced format, but deserialize any format:

#[derive(serde_derive::Serialize, serde_derive::Deserialize)]
struct StructB {
    #[serde(serialize_with = "uuid::serde::braced::serialize")]
    id: uuid::Uuid,
}

Functions

deserialize

Deserialize a braced-formatted string as a Uuid.

serialize

Serialize a Uuid as a braced string.