uuid

Struct NonNilUuid

pub struct NonNilUuid(/* private fields */);

A UUID that is guaranteed not to be the nil UUID.

This is useful for representing optional UUIDs more efficiently, as Option<NonNilUuid> takes up the same space as Uuid.

Note that Uuids created by the following methods are guaranteed to be non-nil:

[Uuid::new_v1]

[Uuid::now_v1]

[Uuid::new_v3]

[Uuid::new_v4]

[Uuid::new_v5]

[Uuid::new_v6]

[Uuid::now_v6]

[Uuid::new_v7]

[Uuid::now_v7]

[Uuid::new_v8]

ABI

The NonNilUuid type does not yet have a stable ABI. Its representation or alignment may change. It is currently only guaranteed that NonNilUuid and Option<NonNilUuid> are the same size as Uuid.

Implementations

impl NonNilUuid

pub const fn new(uuid: Uuid) -> Option<Self>

Creates a non-nil UUID if the value is non-nil.

pub const unsafe fn new_unchecked(uuid: Uuid) -> Self

Creates a non-nil without checking whether the value is non-nil. This results in undefined behavior if the value is nil.

Safety

The value must not be nil.

pub const fn get(self) -> Uuid

Get the underlying Uuid value.