Struct uuid::UuidV1Context

pub struct UuidV1Context { /* fields omitted */ }

A stateful context for the v1 generator to help ensure process-wide uniqueness

Methods

impl UuidV1Context

pub fn new(count: u16) -> UuidV1Context

Creates a thread-safe, internally mutable context to help ensure uniqueness

This is a context which can be shared across threads. It maintains an internal counter that is incremented at every request, the value ends up in the clock_seq portion of the V1 uuid (the fourth group). This will improve the probability that the UUID is unique across the process.