Trait uuid::v1::ClockSequence

pub trait ClockSequence {
    fn generate_sequence(&self, seconds: u64, subsec_nanos: u32) -> u16;
}

A trait that abstracts over generation of UUID v1 “Clock Sequence” values.

Required methods

fn generate_sequence(&self, seconds: u64, subsec_nanos: u32) -> u16

Return a 16-bit number that will be used as the “clock sequence” in the UUID. The number must be different if the time has changed since the last time a clock sequence was requested.