Crate uuid

Generate and parse UUIDs.

Provides support for Universally Unique Identifiers (UUIDs). A UUID is a unique 128-bit number, stored as 16 octets. UUIDs are used to assign unique identifiers to entities without requiring a central allocating authority.

They are particularly useful in distributed systems, though can be used in disparate areas, such as databases and network protocols. Typically a UUID is displayed in a readable string form as a sequence of hexadecimal digits, separated into groups by hyphens.

The uniqueness property is not strictly guaranteed, however for all practical purposes, it can be assumed that an unintentional collision would be extremely unlikely.

Dependencies

By default, this crate depends on nothing but std and cannot generate [Uuid]s. You need to enable the following Cargo features to enable various pieces of functionality:

v1 - adds the Uuid::new_v1 function and the ability to create a V1 using an implementation of uuid::v1::ClockSequence (usually uuid::v1::Context) and a timestamp from time::timespec.

v3 - adds the Uuid::new_v3 function and the ability to create a V3 UUID based on the MD5 hash of some data.

v4 - adds the Uuid::new_v4 function and the ability to randomly generate a Uuid.

v5 - adds the Uuid::new_v5 function and the ability to create a V5 UUID based on the SHA1 hash of some data.

serde - adds the ability to serialize and deserialize a Uuid using the serde crate.

You need to enable one of the following Cargo features together with v3, v4 or v5 feature if you're targeting wasm32 architecture:

stdweb - enables support for OsRng on wasm32-unknown-unknown via stdweb combined with cargo-web

wasm-bindgen - wasm-bindgen enables support for OsRng on wasm32-unknown-unknown via wasm-bindgen

By default, uuid can be depended on with:

[dependencies]
uuid = "0.7"

To activate various features, use syntax like:

[dependencies]
uuid = { version = "0.7", features = ["serde", "v4"] }

You can disable default features with:

[dependencies]
uuid = { version = "0.7", default-features = false }

Examples

To parse a UUID given in the simple format and print it as a urn:

use uuid::Uuid;

fn main() {
    let my_uuid =
        Uuid::parse_str("936DA01F9ABD4d9d80C702AF85C822A8").unwrap();
    println!("{}", my_uuid.to_urn());
}

To create a new random (V4) UUID and print it out in hexadecimal form:

// Note that this requires the `v4` feature enabled in the uuid crate.

use uuid::Uuid;

fn main() {
    let my_uuid = Uuid::new_v4();
    println!("{}", my_uuid);
}

Strings

Examples of string representations:

simple: 936DA01F9ABD4d9d80C702AF85C822A8

hyphenated: 550e8400-e29b-41d4-a716-446655440000

urn: urn:uuid:F9168C5E-CEB2-4faa-B6BF-329BF39FA1E4

References

Wikipedia: Universally Unique Identifier

RFC4122: A Universally Unique IDentifier (UUID) URN Namespace

Re-exports

pub use builder::Builder;

Modules

adapter
Adapters for various formats for Uuids

builder
A Builder type for Uuids.

parser
Uuid parsing constructs and utilities.

prelude
The uuid prelude.

Structs

BytesError
The error that can occur when creating a Uuid.

Uuid
A Universally Unique Identifier (UUID).

Enums

Error
A general error that can occur when handling Uuids.

Variant
The reserved variants of UUIDs.

Version
The version of the UUID, denoting the generating algorithm.

Type Definitions

Bytes
A 128-bit (16 byte) buffer containing the ID.