uuid

Module fmt

Adapters for alternative string formats.

Macros

impl_fmt_from 🔒

impl_fmt_traits 🔒

Structs

Braced

Format a Uuid as a braced hyphenated string, like {67e55044-10b1-426f-9247-bb680e5fe0c8}.

Hyphenated

Format a Uuid as a hyphenated string, like 67e55044-10b1-426f-9247-bb680e5fe0c8.

Simple

Format a Uuid as a simple string, like 67e5504410b1426f9247bb680e5fe0c8.

Urn

Format a Uuid as a URN string, like urn:uuid:67e55044-10b1-426f-9247-bb680e5fe0c8.

Constants

LOWER 🔒

UPPER 🔒

Functions

encode_braced 🔒

encode_hyphenated 🔒

encode_simple 🔒

encode_urn 🔒

format_hyphenated 🔒

format_simple 🔒