uuid

Module serde

Adapters for alternative serde formats.

This module contains adapters you can use with #[serde(with)] to change the way a Uuid is serialized and deserialized.

Modules

braced

Serialize from a Uuid as a uuid::fmt::Braced

compact

Serialize a Uuid as a [u8; 16].

simple

Serialize from a Uuid as a uuid::fmt::Simple

urn

Serialize from a Uuid as a uuid::fmt::Urn