Trait uuid::timestamp::ClockSequence

pub trait ClockSequence {
    type Output;

    fn generate_sequence(&self, seconds: u64, subsec_nanos: u32) -> Self::Output;
}

A counter that can be used by version 1 and version 6 UUIDs to support the uniqueness of timestamps.

References

Clock Sequence in RFC4122

Required Associated Types

type Output

The type of sequence returned by this counter.

Required Methods

fn generate_sequence(&self, seconds: u64, subsec_nanos: u32) -> Self::Output

Get the next value in the sequence to feed into a timestamp.

This method will be called each time a Timestamp is constructed.